requirements.txt
openai-whisper
ffmpeg-python
sentence-transformers
psycopg2-binary
python-dotenv
aiogram==3.4.1
